Update README.md
This project is about Universal Technologies Group that focuses on scaling up wisdom and excellence in technologies. 

It is Your Technology which will be developed in a collaborative manner. 

It is Our Technologies Group that pushes and transforms the current technology poise and strength. 

It is Her Technology that she codes collaboratively and freely. 

It is His Technology that he will work on with us and hers. 

It is Technology of the Time, it is a time technology. Let's move faster and smarter with time